+++ src/writer/mod.rs
@@ -307,7 +307,7 @@ impl<T> Ext for T {
 /// [`Writer`]s, the code won't compile because of the wrong pipelining order.
 ///
 /// ```rust,compile_fail
-/// # use cucumber::{writer, World, WorldInit as _, WriterExt as _};
+/// # use cucumber::{writer, World, WriterExt as _};
 /// #
 /// # #[derive(Debug, Default, World)]
 /// # struct MyWorld;
@@ -328,7 +328,7 @@ impl<T> Ext for T {
 /// ```
 ///
 /// ```rust,compile_fail
-/// # use cucumber::{writer, World, WorldInit as _, WriterExt as _};
+/// # use cucumber::{writer, World, WriterExt as _};
 /// #
 /// # #[derive(Debug, Default, World)]
 /// # struct MyWorld;
@@ -351,7 +351,7 @@ impl<T> Ext for T {
 /// ```rust
 /// # use std::panic::AssertUnwindSafe;
 /// #
-/// # use cucumber::{writer, World, WorldInit as _, WriterExt as _};
+/// # use cucumber::{writer, World, WriterExt as _};
 /// # use futures::FutureExt as _;
 /// #
 /// # #[derive(Debug, Default, World)]